requirements.txt
PyQt6>=6.7,<7.0
holidays>=0.74
openpyxl>=3.1
reportlab>=4.2
PyYAML>=6.0
